fix (GastoEntity): altera preco para double

+++ src/Api.Domain/Entities/GastoEntity.cs
@@ -9,7 +9,7 @@ public class GastoEntity
     public string Usuario { get; set; } = "";
     [Key]
     public string Tipo { get; set; } = "";
-    public decimal? Valor { get; set; }
+    public double? Valor { get; set; }
     public bool Pago { get; set; }
     public DateTime? DataMax { get; set; }
 }
